refactor: moved (obvious) files into functional project structure
