+++ src/components/Banner.js
@@ -25,10 +25,10 @@ import AddLogbookDialog from './AddLogbookDialog';
 import AddTagDialog from './AddTagDialog';
 import LoginDialog from './LoginDialog';
 import LogoutDialog from './LogoutDialog';
-import checkSession from './session-check';
+import checkSession from '../api/session-check';
 // Need axios for back-end access as the "fetch" API does not support CORS cookies.
 import axios from 'axios';
-import packageInfo from '../package.json';
+import packageInfo from '../../package.json';
 import Row from 'react-bootstrap/Row';
 
 /**
