+++ src/components/EmbedImageDialog.js
