+++ src/components/EntryEditor.js
@@ -26,13 +26,13 @@ import { FaPlus } from 'react-icons/fa';
 import { withRouter } from 'react-router-dom';
 import { v4 as uuidv4 } from 'uuid';
 import Attachment from './Attachment.js';
-import customization from './customization';
+import customization from '../utils/customization';
 import EmbedImageDialog from './EmbedImageDialog';
 import OlogAttachment from './OlogAttachment';
 import PropertyEditor from './PropertyEditor';
 import PropertySelector from './PropertySelector';
-import checkSession from './session-check';
-import {removeImageMarkup, ologClientInfoHeader } from './utils';
+import checkSession from '../api/session-check';
+import {removeImageMarkup, ologClientInfoHeader } from '../utils/utils';
 import HtmlPreview from './HtmlPreview';
 import LoadingOverlay from 'react-loading-overlay';
 import Select from 'react-select';
