+++ src/components/Filters.js
@@ -20,13 +20,13 @@ import Logbooks from './Logbooks';
 import Tags from './Tags';
 import { FaCalendarAlt } from "react-icons/fa";
 import Container from 'react-bootstrap/Container';
-import {setSearchParam, removeSearchParam, dateToString} from './utils';
+import {setSearchParam, removeSearchParam, dateToString} from '../utils/utils';
 import Form from 'react-bootstrap/Form';
 import Table from 'react-bootstrap/Table';
 import DateTimePicker from 'react-datetime-picker';
 import Modal from 'react-bootstrap/Modal';
 import Button from 'react-bootstrap/Button';
-import customization from './customization';
+import customization from '../utils/customization';
 
 /**
  * Component holding search criteria elements, i.e.
