+++ src/components/Properties.js
