+++ src/components/SearchStringHelpDialog.js
