+++ src/utils/utils.js
@@ -17,7 +17,7 @@
  */
 import moment from 'moment';
 import { v4 as uuidv4 } from 'uuid';
-import packageInfo from '../package.json';
+import packageInfo from '../../package.json';
 
  const shortTimeFormat = 'HH:mm';
  const shortDateFormat = 'YYYY-MM-DD';
